Check if binary exists then delete on autoinstall

+++ autoinstall.sh
@@ -1,6 +1,9 @@
 javac UuidGenerator.java
 mkdir -p /usr/local/lib/uuid
 cp UuidGenerator.class /usr/local/lib/uuid
-rm /usr/local/bin/uuid
+if [ -e /usr/local/bin/uuid ]
+then 
+	rm /usr/local/bin/uuid
+fi
 echo 'java --class-path=/usr/local/lib/uuid UuidGenerator $1' >> /usr/local/bin/uuid
 chmod a+x /usr/local/bin/uuid
